Working with a project: adding feed logic; setting up the project structure
