+++ src/main/java/com/example/socialmediaapi/contracts/ImageService.java
@@ -1,0 +1,8 @@
+package com.example.socialmediaapi.contracts;
+
+public interface ImageService {
+
+    byte[] compressBytes(byte[] data);
+    byte[] decompressBytes(byte[] data);
+
+}
