+++ src/main/java/com/example/socialmediaapi/contracts/RoleService.java
@@ -1,0 +1,9 @@
+package com.example.socialmediaapi.contracts;
+
+import com.example.socialmediaapi.core.entities.Role;
+
+public interface RoleService {
+
+    Role findRoleByName(String roleName);
+
+}
