+++ src/main/java/com/example/socialmediaapi/contracts/SubscriptionService.java
@@ -1,0 +1,12 @@
+package com.example.socialmediaapi.contracts;
+
+import java.util.UUID;
+
+public interface SubscriptionService {
+
+    void subscribe(UUID targetId, String followerUsername);
+    void acceptFriendship(UUID targetId, String acceptingUsername);
+    void unsubscribe(UUID targetId, String initiatorUsername);
+    void unfriend(UUID targetId, String initiatorUsername);
+
+}
