+++ src/main/java/com/example/socialmediaapi/converters/PostOutConverter.java
@@ -1,0 +1,24 @@
+package com.example.socialmediaapi.converters;
+
+import com.example.socialmediaapi.contracts.ImageService;
+import com.example.socialmediaapi.dto.responces.PostDtoOut;
+import com.example.socialmediaapi.core.entities.Post;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class PostOutConverter {
+
+    private final ImageService imageService;
+
+    public PostDtoOut entityToDto(Post post) {
+        PostDtoOut postDtoOut = new PostDtoOut();
+        postDtoOut.setTitle(post.getTitle());
+        postDtoOut.setText(post.getText());
+        postDtoOut.setPublisherUsername(post.getUser().getUsername());
+        postDtoOut.setImage(imageService.decompressBytes(post.getImage()));
+        return postDtoOut;
+    }
+
+}
