+++ src/main/java/com/example/socialmediaapi/core/repositories/RoleRepository.java
@@ -1,0 +1,10 @@
+package com.example.socialmediaapi.core.repositories;
+
+import com.example.socialmediaapi.core.entities.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RoleRepository extends JpaRepository<Role, Long> {
+    Role findByName(String name);
+}
