+++ src/main/java/com/example/socialmediaapi/core/services/ImageServiceImpl.java
@@ -1,0 +1,55 @@
+package com.example.socialmediaapi.core.services;
+
+import com.example.socialmediaapi.contracts.ImageService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.util.zip.DataFormatException;
+import java.util.zip.Deflater;
+import java.util.zip.Inflater;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class ImageServiceImpl implements ImageService {
+
+    @Override
+    public byte[] compressBytes(byte[] data) {
+        Deflater deflater = new Deflater();
+        deflater.setInput(data);
+        deflater.finish();
+        ByteArrayOutputStream outputStream = new ByteArrayOutputStream(data.length);
+        byte[] buffer = new byte[1024];
+        while (!deflater.finished()) {
+            int count = deflater.deflate(buffer);
+            outputStream.write(buffer, 0, count);
+        }
+        try {
+            outputStream.close();
+        } catch (IOException e) {
+            log.error("Cannot compress Bytes");
+        }
+        return outputStream.toByteArray();
+    }
+
+    @Override
+    public byte[] decompressBytes(byte[] data) {
+        Inflater inflater = new Inflater();
+        inflater.setInput(data);
+        ByteArrayOutputStream outputStream = new ByteArrayOutputStream(data.length);
+        byte[] buffer = new byte[1024];
+        try {
+            while (!inflater.finished()) {
+                int count = inflater.inflate(buffer);
+                outputStream.write(buffer, 0, count);
+            }
+            outputStream.close();
+        } catch (IOException | DataFormatException e) {
+            log.error("Cannot decompress Bytes");
+        }
+        return outputStream.toByteArray();
+    }
+}
