--- src/main/java/com/example/socialmediaapi/dto/PostDto.java
@@ -1,18 +1,0 @@
-package com.example.socialmediaapi.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class PostDto {
-
-    private UUID id;
-    private String title;
-    private String text;
-
-}
