+++ src/main/java/com/example/socialmediaapi/dto/requests/PostDtoIn.java
@@ -1,0 +1,14 @@
+package com.example.socialmediaapi.dto.requests;
+
+import lombok.Data;
+
+import java.util.UUID;
+
+@Data
+public class PostDtoIn {
+
+    private UUID id;
+    private String title;
+    private String text;
+
+}
