+++ src/main/java/com/example/socialmediaapi/dto/responces/PostDtoOut.java
@@ -1,0 +1,13 @@
+package com.example.socialmediaapi.dto.responces;
+
+import lombok.Data;
+
+@Data
+public class PostDtoOut {
+
+    private String publisherUsername;
+    private String title;
+    private String text;
+    private byte[] image;
+
+}
