--- src/main/java/com/example/socialmediaapi/entities/Message.java
@@ -1,35 +1,0 @@
-package com.example.socialmediaapi.entities;
-
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "messages")
-public class Message {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String text;
-
-    @ManyToOne
-    @JoinColumn(name = "sender")
-    private User sender;
-
-    @ManyToOne
-    @JoinColumn(name = "addressee")
-    private User addressee;
-
-    @CreationTimestamp
-    @Column(name = "created_at")
-    private LocalDateTime createdAt;
-}
