--- src/main/java/com/example/socialmediaapi/entities/Subscription.java
@@ -1,29 +1,0 @@
-package com.example.socialmediaapi.entities;
-
-import lombok.*;
-
-import javax.persistence.*;
-
-@Entity
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "subscriptions")
-public class Subscription {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "follower")
-    private User follower;
-
-    @ManyToOne
-    @JoinColumn(name = "subscription_target")
-    private User subscriptionTarget;
-
-    private boolean isFriends;
-}
