--- src/main/java/com/example/socialmediaapi/repositories/PostRepository.java
@@ -1,15 +1,0 @@
-package com.example.socialmediaapi.repositories;
-
-import com.example.socialmediaapi.entities.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.UUID;
-
-@Repository
-public interface PostRepository extends JpaRepository<Post, UUID> {
-
-    List<Post> findAllByUserId(UUID userId);
-
-}
