--- src/main/java/com/example/socialmediaapi/repositories/RoleRepository.java
@@ -1,10 +1,0 @@
-package com.example.socialmediaapi.repositories;
-
-import com.example.socialmediaapi.entities.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
-    Role findByName(String name);
-}
